Major COLLADA overhaul, added a processor for "Polygons" type data. Still has some work to be done on it though!

+++ src/imi/scene/polygonmodel/parts/PMeshMaterialCombo.java
@@ -211,7 +211,7 @@ public class PMeshMaterialCombo
                 float avg   =  m_Material1.getShininess();
                 avg         += m_Material2.getShininess();
                 avg         *= 0.5f;
-                m_MaterialCombo.setShininess(avg);
+                m_MaterialCombo.setShininess((int)avg);
                 break;
             }
         }
